requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coverage==7.6.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.1
pytest-cov==5.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.8.0
requests==2.32.3
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
